refactor: auth service type 관련 코드 type.ts로 분리
